vehicle-damage-ai/requirements.txt
# Core API clients
anthropic>=0.40.0
openai>=1.50.0
google-generativeai>=0.8.0

# Data processing
pandas>=2.1.0
numpy>=1.26.0

# Visualization
matplotlib>=3.8.0
plotly>=5.22.0
kaleido>=0.2.1

# Image handling
Pillow>=10.2.0

# Optional: Streamlit demo UI
streamlit>=1.39.0

# Optional: MCP integration (Week 11-12 only)
# mcp>=1.0.0
